pyproject.toml
[project]
name = "capstone-text-digest"
version = "1.0.0"
description = "Text summarization tool with style adaptation using Groq AI and Streamlit"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "streamlit>=1.35.0",
    "groq>=0.9.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "striprtf>=0.0.26",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "fpdf2>=2.7.9",
    "gtts>=2.5.0",
    "better-profanity>=0.7.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
